refactor: pure reorganize of types
